+++ token-swap/program/src/constraints.rs
@@ -1,8 +1,8 @@
 //! Various constraints as required for production environments
 
 use crate::{
+    curve::base::{CurveType, SwapCurve},
     curve::{constant_product::ConstantProductCurve, flat::FlatCurve},
-    curve::shared::{CurveType, SwapCurve},
     error::SwapError,
 };
 
@@ -102,7 +102,7 @@ pub const FEE_CONSTRAINTS: Option<FeeConstraints> = {
 mod tests {
     use super::*;
 
-    use crate::curve::{constant_product::ConstantProductCurve, shared::CurveType};
+    use crate::curve::{base::CurveType, constant_product::ConstantProductCurve};
 
     #[test]
     fn validate_fees() {
