+++ token-swap/program/src/curve/mod.rs
@@ -1,5 +1,6 @@
 //! ...
 
+pub mod base;
+pub mod calculator;
 pub mod constant_product;
 pub mod flat;
-pub mod shared;
